requirements.txt
fastapi
uvicorn
google-generativeai
firebase-admin
python-multipart
openpyxl
python-dotenv
httpx
pandas
